renaming some files
