+++ app/src/main/java/com/arturkida/popularmovies_kotlin/ui/movies/MoviesViewModel.kt
@@ -1,4 +1,4 @@
-package com.arturkida.popularmovies_kotlin.ui.popular
+package com.arturkida.popularmovies_kotlin.ui.movies
 
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
