+++ app/src/main/res/layout/movies_activity.xml
@@ -5,4 +5,4 @@
         android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.popular.MoviesActivity"/>
+        tools:context=".ui.movies.MoviesActivity"/>
